production/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
openpyxl>=3.0.0

# ML models
xgboost>=1.7.0
catboost>=1.1.0

# Hyperparameter optimization
optuna>=3.0.0

# Configuration
pyyaml>=6.0

# API
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
joblib>=1.2.0
python-dateutil>=2.8.0

# Optional: Production server
gunicorn>=21.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=6.0.0
